requirements.txt
crewai
crewai_tools
python-dotenv
langchain-google-genai
google-generativeai
fastapi
uvicorn[standard]
python-multipart
